Moved utility functions their respective files for organization.
